Copyright modification
Part 4 (in this case, notably 4.a, 4.b and 4.d) of the Apache License is there for a reason.
Keep the license in tact. If you would like to license your additions and/or modifications, clearly outline any changes you have made in the changed files and include your license there. Removing or altering the original license is copyright infringement.
I have sent a pull request only for the README for the sake of notifying you, but the headers of index.html, main.css and main.js have all been modified, had their license removed, and need to be fixed.